[add] cf#543, #544, abc120, wupc2019

+++ Source/arc085/arc085_a.cpp
@@ -18,7 +18,7 @@
 
 using namespace std;
  
-typedef int64_t i64;
+typedef long long i64;
 typedef pair<i64, i64> P;
 
 template<class T>
